inventory: Fix build_response_switch()
1. app/v1/endpoints/lan_fabric/rest/control/fabrics/inventory/switches_by_fabric_get.py

- build_response_switch()
  - Move to shared location.
  - Import from shared location.

2. app/v1/endpoints/lan_fabric/rest/control/fabrics/inventory/internal_inventory_get.py

- build_response_switch()
  - Remove and import from, shared location.

3. app/v1/endpoints/lan_fabric/rest/control/fabrics/inventory/common.py

- build_response_switch()
  - Moved from switches_by_fabric_get.py
  - Both switchRole and switchRoleEnum were missing.  Added them.
